requirements.txt
PyJWT==2.6.0
cryptography==41.0.2
requests==2.31.0
python-jose==3.2.0
